+++ apps/server/src/modules/tool/service/validation/common-tool-validation.service.spec.ts
@@ -1,0 +1,199 @@
+import {
+	customParameterDOFactory,
+	externalToolDOFactory,
+} from '@shared/testing/factory/domainobject/external-tool.factory';
+import { Test, TestingModule } from '@nestjs/testing';
+import { createMock, DeepMocked } from '@golevelup/ts-jest';
+import { CustomParameterDO, ExternalToolDO } from '@shared/domain/domainobject/external-tool';
+import { UnprocessableEntityException } from '@nestjs/common';
+import { ExternalToolService } from '../external-tool.service';
+import { CommonToolValidationService } from './common-tool-validation.service';
+
+describe('CommonToolValidationService', () => {
+	let module: TestingModule;
+	let service: CommonToolValidationService;
+
+	let externalToolService: DeepMocked<ExternalToolService>;
+
+	beforeAll(async () => {
+		module = await Test.createTestingModule({
+			providers: [
+				CommonToolValidationService,
+				{
+					provide: ExternalToolService,
+					useValue: createMock<ExternalToolService>(),
+				},
+			],
+		}).compile();
+
+		service = module.get(CommonToolValidationService);
+		externalToolService = module.get(ExternalToolService);
+	});
+
+	afterAll(async () => {
+		await module.close();
+	});
+
+	afterEach(() => {
+		jest.clearAllMocks();
+	});
+
+	describe('validateCommon is called', () => {
+		describe('when tool is valid', () => {
+			it('should throw no excpetion', async () => {
+				const externalToolDO: ExternalToolDO = externalToolDOFactory.build({
+					name: undefined,
+					parameters: [
+						customParameterDOFactory.build({ name: 'sameKey' }),
+						customParameterDOFactory.build({ name: 'notSameKey' }),
+					],
+				});
+
+				const func = () => service.validateCommon(externalToolDO);
+
+				await expect(func()).resolves.not.toThrow();
+			});
+		});
+
+		describe('when name already exists', () => {
+			it('should find itself', async () => {
+				const externalToolDO: ExternalToolDO = externalToolDOFactory.buildWithId();
+				externalToolService.findExternalToolByName.mockResolvedValue(externalToolDO);
+
+				const result: Promise<void> = service.validateCommon(externalToolDO);
+
+				await expect(result).resolves.not.toThrow();
+			});
+
+			it('should find a tool with the same name', async () => {
+				const externalToolDO: ExternalToolDO = externalToolDOFactory.build({ name: 'sameName' });
+				const existingExternalToolDO: ExternalToolDO = externalToolDOFactory.buildWithId({ name: 'sameName' });
+				externalToolService.findExternalToolByName.mockResolvedValue(existingExternalToolDO);
+
+				const result: Promise<void> = service.validateCommon(externalToolDO);
+
+				await expect(result).rejects.toThrow(
+					new UnprocessableEntityException(`The tool name "${externalToolDO.name}" is already used.`)
+				);
+			});
+		});
+
+		describe('when there are duplicate attributes', () => {
+			it('should fail for two equal parameters', async () => {
+				const externalToolDO: ExternalToolDO = externalToolDOFactory.build({
+					parameters: [
+						customParameterDOFactory.build({ name: 'paramEqual' }),
+						customParameterDOFactory.build({ name: 'paramEqual' }),
+					],
+				});
+				externalToolService.findExternalToolByName.mockResolvedValue(null);
+
+				const func = () => service.validateCommon(externalToolDO);
+
+				await expect(func()).rejects.toThrow(
+					new UnprocessableEntityException(
+						`The tool ${externalToolDO.name} contains multiple of the same custom parameters.`
+					)
+				);
+			});
+
+			it('should fail for names that only differ in capitalisation', async () => {
+				const externalToolDO: ExternalToolDO = externalToolDOFactory.build({
+					parameters: [
+						customParameterDOFactory.build({ name: 'param1CaseSensitive' }),
+						customParameterDOFactory.build({ name: 'Param1casesensitive' }),
+					],
+				});
+				externalToolService.findExternalToolByName.mockResolvedValue(null);
+
+				const result: Promise<void> = service.validateCommon(externalToolDO);
+
+				await expect(result).rejects.toThrow(
+					new UnprocessableEntityException(
+						`The tool ${externalToolDO.name} contains multiple of the same custom parameters.`
+					)
+				);
+			});
+		});
+
+		describe('when regex is invalid', () => {
+			it('throw when external tools has a faulty regular expression', async () => {
+				const externalToolDO: ExternalToolDO = externalToolDOFactory.withCustomParameters(1, { regex: '[' }).build();
+				externalToolService.findExternalToolByName.mockResolvedValue(null);
+
+				const func = () => service.validateCommon(externalToolDO);
+
+				await expect(func()).rejects.toThrow(
+					new UnprocessableEntityException(
+						`A custom Parameter of the tool ${externalToolDO.name} has wrong regex attribute.`
+					)
+				);
+			});
+		});
+
+		describe('when regex is set but regex comment is missing', () => {
+			it('should throw exception', async () => {
+				const externalToolDO: ExternalToolDO = externalToolDOFactory.withCustomParameters(1, { regex: '.' }).build();
+				const result: Promise<void> = service.validateCommon(externalToolDO);
+
+				await expect(result).rejects.toThrow(
+					new UnprocessableEntityException(
+						`The "${(externalToolDO.parameters as CustomParameterDO[])[0].name}" parameter is missing a regex comment.`
+					)
+				);
+			});
+		});
+	});
+
+	describe('isClientIdUnique is called', () => {
+		describe('when config is not of type oauth2', () => {
+			it('should return without error', async () => {
+				const externalToolDO: ExternalToolDO = externalToolDOFactory.build();
+
+				const func = () => service.isClientIdUnique(externalToolDO);
+
+				await expect(func()).resolves.not.toThrow();
+			});
+		});
+
+		describe('when config is of type oauth2', () => {
+			it('should call the service', async () => {
+				const expectedClientId = 'clientId';
+				const externalToolDO: ExternalToolDO = externalToolDOFactory
+					.withOauth2Config({ clientId: expectedClientId })
+					.build();
+
+				await service.isClientIdUnique(externalToolDO);
+
+				expect(externalToolService.findExternalToolByOAuth2ConfigClientId).toHaveBeenCalledWith(expectedClientId);
+			});
+		});
+
+		describe('when tool name does not have changed', () => {
+			it('should return true', async () => {
+				const externalToolDO: ExternalToolDO = externalToolDOFactory.withOauth2Config().buildWithId();
+				externalToolService.findExternalToolByOAuth2ConfigClientId.mockResolvedValue(externalToolDO);
+
+				const result: boolean = await service.isClientIdUnique(externalToolDO);
+
+				expect(result).toBeTruthy();
+			});
+		});
+
+		describe('when another tool with the name already exists', () => {
+			it('should return false', async () => {
+				const externalToolDO: ExternalToolDO = externalToolDOFactory
+					.withOauth2Config({ clientId: 'sameClientId' })
+					.build();
+				const existingExternalToolDO: ExternalToolDO = externalToolDOFactory
+					.withOauth2Config({ clientId: 'sameClientId' })
+					.buildWithId();
+				externalToolService.findExternalToolByOAuth2ConfigClientId.mockResolvedValue(existingExternalToolDO);
+
+				const result: boolean = await service.isClientIdUnique(externalToolDO);
+
+				expect(result).toBeFalsy();
+			});
+		});
+	});
+});
